q2_chat_with_lecture/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
openai>=1.6.1
chromadb==0.4.18
sentence-transformers==2.2.2
moviepy==1.0.3
pydub==0.25.1
whisper==1.1.10
python-dotenv==1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 